requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-google-genai
langchain-huggingface
google-generativeai
pandas
pypdf
docx2txt
chromadb
sentence-transformers
python-dotenv
